Bryan: Nit fixes + Additions to About Me
Directory name change/Added new CSS file for AboutMe section; Also
changed paths to use @/personal/bryan; Gave better names to #position
css ids; Added interests section to AboutMeText
