+++ js/src/Routes.tsx
@@ -3,7 +3,7 @@ import { BryanPage } from './personal/bryan/Bryan.page.tsx'
 import { MainPage as BryanMainPage } from './personal/bryan/Main/Main.page.tsx'
 import { ProjectsPage as BryanProjectsPage } from './personal/bryan/Projects/Projects.page.tsx'
 import { FunPage as BryanFunPage } from './personal/bryan/Fun/Fun.page.tsx'
-import { AboutMePage as BryanAboutMePage } from './personal/bryan/About Me/AboutMe.page.tsx'
+import { AboutMePage as BryanAboutMePage } from '@/personal/bryan/AboutMe/AboutMe.page.tsx'
 import { HaokingPage } from './personal/haoking/Haoking.page.tsx'
 import { HenryPage } from './personal/_henry/Henry.page.tsx'
 import { HomePage } from './pages/Home.page'
