+++ js/src/personal/bryan/Bryan.module.css
@@ -18,26 +18,11 @@
   }
 }
 
-#position1 {
-  position: absolute;
-  top: 20rem;
-  left: 5rem;
-}
-
 .header {
     height: rem(56px);
     margin-bottom: rem(60px);
     background-color: var(--mantine-color-body);
     border-bottom: rem(1px) solid light-dark(var(--mantine-color-gray-3), var(--mantine-color-dark-4));
-}
-
-.playlist{
-  position: absolute;
-  border: none;
-  top: 20rem;
-  width: 300px;
-  height: 450px;
-  right: 5rem;
 }
 
 .inner {
