+++ js/src/personal/bryan/Bryan.page.tsx
@@ -2,8 +2,8 @@ import { MantineProvider } from '@mantine/core'
 import { Outlet } from 'react-router-dom'
 import { theme } from '@/personal/bryan/theme.ts'
 import { NavBar } from '@/personal/bryan/Frame/NavBar.tsx'
-import classes from './Bryan.module.css'
-import { Socials } from './Frame/Socials.tsx'
+import classes from '@/personal/bryan/Bryan.module.css'
+import { Socials } from '@/personal/bryan/Frame/Socials.tsx'
 import { TimeDisplay } from '@/personal/bryan/Frame/TimeDisplay.tsx'
 
 export function BryanPage() {
